Desktop/windsurf/CascadeProjects/windsurf-project/mental-health-tracker/backend/requirements.txt
flask==2.3.3
werkzeug==2.3.7
flask-sqlalchemy==3.0.5
flask-migrate==4.1.0
flask-cors==4.0.0
flask-jwt-extended==4.7.1
python-dotenv==1.0.0
SQLAlchemy==2.0.23
openai==0.27.8
requests==2.31.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.1.0
google-api-python-client==2.108.0
PyJWT==2.1.0
